Fix bug in backward of interpolate function. The output gradient can be undefined, need to add check for that.
Summary:
In forward we set `ctx->set_materialize_grads(false);`
This means that autograd engine can optimize gradient computations in the cases where the output does not depend on the input by not materializing grad tensors given to backward function. Not materializing means that if will pass undefined tensor.

However, the `interpolate` function has a single output. Thus if the output is not differentiated it would reasonable to expect autograd not to call backward at all. Unfortunetly that seems to be not true, and we still need to check if the output gradient is defined. This behaivour is probably specific to DDP.

Differential Revision: D56937038

fbshipit-source-id: 946597fb524f387ddcfaab64b5c5ee6a57689923