Don't ship internal only files
Summary: This isn't the real fix it's just moving the files into a folder we can exclude in shipit. See D62289505 where we exclude this directory from shipit. This project is exported to open source (Github) so we need to not ship these files -- I didn't realize this when I wrote the tests.

Reviewed By: jermenkoo

Differential Revision: D62289557

fbshipit-source-id: 60afb00fd8a6aa54aabe12a8a3fcc161a65b0495