Fix bug in the autocast wrapper of the edge_grad
Summary: Fixes the autocast wrapper for the edge grad. Most likely due to the copypast, it had the `[0]` before return, which for the autograd wrapper makes sense since the autograd function always returns a list of tensors. But in the autocasty wrapper, we are calling a dispatch wrapper, that directly returns single tensor, and by doing `[0]` we erroneously  return the first element of the batch.

Reviewed By: tsimk

Differential Revision:
D85078237

Privacy Context Container: L900159

fbshipit-source-id: 76c2583cdb95a3d7d086e0e0879ccb4e50fd7b69